Time datatype + version_flag
- Time datatype from `int` to `int64_t` to guarantize the portability in
mobile platforms (see C++ reference for time()).

- `version_flag` changed to be the `isExported` flag directly, saving a
byte in serialized data.

+++ include/megaapi.h
@@ -4277,7 +4277,7 @@ class MegaApi
          *
          * @note A Unix timestamp represents the number of seconds since 00:00 hours, Jan 1, 1970 UTC
          */
-        void exportNode(MegaNode *node, int expireTime, MegaRequestListener *listener = NULL);
+        void exportNode(MegaNode *node, int64_t expireTime, MegaRequestListener *listener = NULL);
 
         /**
          * @brief Stop sharing a file/folder
